refactor: View → ViewModel 로직 분기 리팩토링
- View이름 변경 (RestaurantMapView → PlaceMapView)
- 오타수정
- NameSpace 추가
