+++ BestEats/View/PlaceMap/PlaceMapView.swift
@@ -1,5 +1,5 @@
 //
-//  RecommendationView.swift
+//  PlaceMapView.swift
 //  BestEats
 //
 //  Created by BH on 2024/07/16.
@@ -8,27 +8,15 @@
 import SwiftUI
 import MapKit
 
-enum PlaceInfoError: Error {
-    case emptyData
-    
-    var messageDescription: String {
-        switch self {
-        case .emptyData:
-            return "정보없음"
-        }
-    }
-}
-
-struct RestaurantMapView: View {
+struct PlaceMapView: View {
     
     @Environment(\.dismiss) private var dismiss
-    @StateObject private var vm = RestaurantMapViewModel()
+    @StateObject private var vm = PlaceMapViewModel()
     @State private var isPresentedSheet: Bool = false
-    @State private var nearestPlace: V2.Local.Search.Keyword.Response.PlaceInfo? = nil
-    @State private var selectedPlace: V2.Local.Search.Keyword.Response.PlaceInfo?
     @State var foodType: FoodType = .cafe
     
     var body: some View {
+        NavigationStack {
         ZStack {
             map
             
@@ -40,20 +28,13 @@ struct RestaurantMapView: View {
             .padding(.horizontal)
             .padding(.vertical, 24)
         }
+    }
         .onAppear {
             vm.getCurrentLocation()
             self.isPresentedSheet = true
         }
-        .onChange(of: vm.nearRestaurants) { newPlaces in
-            if let place = newPlaces.first {
-                self.nearestPlace = place
-                zoomToLocation(location: place.coordinate)
-            }
-        }
-        .onChange(of: selectedPlace) { newPlace in
-            if let selectedPlace = vm.nearRestaurants.first(where: { $0.id == newPlace!.id }) {
-                zoomToLocation(location: selectedPlace.coordinate)
-            }
+        .onChange(of: vm.place) { newPlace in
+            zoomToLocation(location: newPlace!.coordinate)
         }
         .onChange(of: foodType) { newFoodType in
             Task { await vm.fetchNearRestaurant(foodType: newFoodType) }
@@ -67,52 +48,52 @@ struct RestaurantMapView: View {
     
     // MARK: - Function
     
-    private func setInfoNearPlaceCount() -> String {
-        let placeCount: Int = vm.nearRestaurants.count
-        return "근처에 \(placeCount)개의 맛집이 있어요!"
-    }
-    
-    private func setPlaceName() -> String {
-        guard let place = selectedPlace else {
-            return self.nearestPlace?.placeName ?? PlaceInfoError.emptyData.messageDescription
-        }
-        return place.placeName
-    }
-    
-    private func setCategoryName() -> String {
-        
-        guard let place = selectedPlace,
-              let categoryName = place.categoryName else {
-            return self.nearestPlace?.categoryName ?? PlaceInfoError.emptyData.messageDescription
-        }
-        return categoryName
-    }
-    
-    private func setAddressName() -> String {
-        guard let place = selectedPlace,
-              let roadAddress = place.roadAddressName else {
-            return self.nearestPlace?.roadAddressName ?? PlaceInfoError.emptyData.messageDescription
-        }
-        return place.addressName ?? roadAddress
-    }
-    
-    private func setDistance() -> String {
-        var distance: String?
-        
-        if let place = selectedPlace {
-            distance = place.distance
-        } else {
-            distance = nearestPlace?.distance
-        }
-        
-        guard let validDistance = distance,
-              let distanceValue = Double(validDistance) else {
-            return PlaceInfoError.emptyData.messageDescription
-        }
-        
-        return distanceValue.convertedDistance()
-    }
-    
+//    private func setInfoNearPlaceCount() -> String {
+//        let placeCount: Int = vm.placeList.count
+//        return "근처에 \(placeCount)개의 맛집이 있어요!"
+//    }
+//    
+//    private func setPlaceName() -> String {
+//        guard let place = selectedPlace else {
+//            return self.nearestPlace?.placeName ?? PlaceInfoError.emptyData.messageDescription
+//        }
+//        return place.placeName
+//    }
+//    
+//    private func setCategoryName() -> String {
+//        
+//        guard let place = selectedPlace,
+//              let categoryName = place.categoryName else {
+//            return self.nearestPlace?.categoryName ?? PlaceInfoError.emptyData.messageDescription
+//        }
+//        return categoryName
+//    }
+//    
+//    private func setAddressName() -> String {
+//        guard let place = selectedPlace,
+//              let roadAddress = place.roadAddressName else {
+//            return self.nearestPlace?.roadAddressName ?? PlaceInfoError.emptyData.messageDescription
+//        }
+//        return place.addressName ?? roadAddress
+//    }
+//    
+//    private func setDistance() -> String {
+//        var distance: String?
+//        
+//        if let place = selectedPlace {
+//            distance = place.distance
+//        } else {
+//            distance = nearestPlace?.distance
+//        }
+//        
+//        guard let validDistance = distance,
+//              let distanceValue = Double(validDistance) else {
+//            return PlaceInfoError.emptyData.messageDescription
+//        }
+//        
+//        return distanceValue.convertedDistance()
+//    }
+//    
     private func zoomToLocation(location: CLLocationCoordinate2D) {
         withAnimation(.easeIn) {
             vm.region.center = location
@@ -122,18 +103,18 @@ struct RestaurantMapView: View {
 
 // MARK: - View
 
-extension RestaurantMapView {
+extension PlaceMapView {
     
     private var map: some View {
         Map(
             coordinateRegion: $vm.region,
             showsUserLocation: true,
             //                userTrackingMode: .constant(.follow),
-            annotationItems: vm.nearRestaurants
+            annotationItems: vm.placeList
         ) { nearPlace in
             MapAnnotation(coordinate: nearPlace.coordinate) {
                 CustomMapMarkerView(
-                    selectedPlace: $selectedPlace,
+                    selectedPlace: $vm.place,
                     place: nearPlace,
                     type: foodType
                 )
@@ -154,7 +135,7 @@ extension RestaurantMapView {
     }
     
     private var searchedPlaceText: some View {
-        Text(setInfoNearPlaceCount())
+        Text(vm.setInfoNearPlaceCount())
             .font(.pretendardBold14)
             .foregroundStyle(.white)
             .padding()
@@ -166,7 +147,7 @@ extension RestaurantMapView {
         Button(action: {
             self.isPresentedSheet = true
         }, label: {
-            Image(systemName: "line.3.horizontal")
+            Image(systemName: Img.Map.Icons.menu)
                 .font(.pretendardBold18)
         })
         .frame(width: 48, height: 48)
@@ -205,14 +186,17 @@ extension RestaurantMapView {
     private var placeInfoAndButtonContainer: some View {
         HStack(spacing: 16) {
             placeInfoContainer
-            placeInfoButtonContainer
+//            if $vm.selectedPlace != nil || $vm.nearestPlace != nil {
+            // TODO: [] 확인 필요
+                placeInfoButtonContainer
+//            }
         }
         .frame(maxHeight: 128)
     }
     
     private var placeInfoContainer: some View {
         VStack(alignment: .leading, spacing: 24) {
-            Text(setPlaceName())
+            Text(vm.setPlaceName())
                 .font(.pretendardBold20)
             
             HStack(alignment: .bottom) {
@@ -228,8 +212,8 @@ extension RestaurantMapView {
     
     private var placeInfoContents: some View {
         VStack(alignment: .leading, spacing: 8) {
-            Text(setCategoryName())
-            Text(setAddressName())
+            Text(vm.setCategoryName())
+            Text(vm.setAddressName())
         }
         .font(.pretendardRegular14)
         .foregroundStyle(.gray)
@@ -241,8 +225,8 @@ extension RestaurantMapView {
             print("길안내는 준비중 입니다")
         } label: {
             HStack {
-                Image(systemName: "location.fill")
-                Text(setDistance())
+                Image(systemName: Img.Map.Icons.location)
+                Text(vm.setDistance())
             }
         }
         .font(.pretendardBold14)
@@ -254,17 +238,22 @@ extension RestaurantMapView {
     
     private var placeInfoButtonContainer: some View {
         VStack {
-            Button("더보기") {
-                // TODO: [] WebView 띄우기
+            NavigationLink {
+                ReuseWebView(
+                    placeName: vm.place?.placeName ?? "",
+                    urlString: vm.place?.placeURL ?? ""
+                )
+            } label: {
+                Text(ETC.Button.more)
+                    .padding()
+                    .frame(width: 88)
+                    .background(.green)
+                    .clipShape(RoundedRectangle(cornerRadius: 16))
             }
-            .padding()
-            .frame(width: 88)
-            .background(.green)
-            .clipShape(RoundedRectangle(cornerRadius: 16))
             
             Spacer()
             
-            Button("예약하기") {
+            Button(PlaceMap.Button.reserve) {
                 // TODO: [] 전화 팝업 띄우기
             }
             .padding()
@@ -279,5 +268,5 @@ extension RestaurantMapView {
 
 #Preview {
     @State var type: FoodType = .cafe
-    return RestaurantMapView()
+    return PlaceMapView()
 }
